requirements.txt
flask==3.1.0
flask-cors>=6.0.1
websocket-client>=1.6.4
requests>=2.32.3
opencv-python>=4.9.0.80
numpy>=1.25.2
reportlab>=4.0.0